requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
supabase==0.0.2
python-dotenv==0.19.0
requests==2.26.0
groq==0.3.0 